revisited static_cast<> for integral types

+++ include/orm/utils/thread.hpp
@@ -28,14 +28,11 @@ namespace Orm::Utils
 
         /*! Set thread name for debugger, -1 for a current thread. */
         [[maybe_unused]]
-        static void nameThreadForDebugging(
-                const QString &threadName,
-                quint64 threadId = static_cast<quint64>(-1));
+        static void nameThreadForDebugging(const QString &threadName,
+                                           quint64 threadId = -1);
         /*! Set thread name for debugger, -1 for a current thread. */
         [[maybe_unused]]
-        static void nameThreadForDebugging(
-                const char *threadName,
-                quint64 threadId = static_cast<quint64>(-1));
+        static void nameThreadForDebugging(const char *threadName, quint64 threadId = -1);
     };
 
 } // namespace Orm::Utils
